Added DAO support

+++ Operativa/src/operativa/model/dao/ItemDAO.java
@@ -7,7 +7,7 @@ public interface ItemDAO extends GenericDAO<Item, Long> {
     public static final String QUERY_MAXBID = "ItemDAO.QUERY_MAXBID";
     public static final String QUERY_MINBID = "ItemDAO.QUERY_MINBID";
  
-    public int getMaxBid(Long itemId);
-    public int getMinBid(Long itemId);
+    public Long getMaxBid(Long itemId);
+    public Long getMinBid(Long itemId);
  
 }
